@@ -35,44 +35,44 @@
     "svelte": "^4.0.0"
   },
   "devDependencies": {
-    "@fontsource/roboto": "^5.0.13",
-    "@material/material-color-utilities": "^0.2.7",
-    "@sveltejs/adapter-auto": "^3.2.1",
-    "@sveltejs/kit": "^2.5.10",
-    "@sveltejs/package": "^2.3.1",
-    "@sveltejs/vite-plugin-svelte": "^3.1.0",
-    "@types/node": "^20.12.12",
-    "@types/prettier": "^2.7.3",
+    "@fontsource/roboto": "^5.1.0",
+    "@material/material-color-utilities": "^0.3.0",
+    "@sveltejs/adapter-auto": "^3.2.5",
+    "@sveltejs/kit": "^2.5.28",
+    "@sveltejs/package": "^2.3.5",
+    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@types/node": "^22.6.1",
+    "@types/prettier": "^3.0.0",
     "@types/prismjs": "^1.26.4",
-    "@typescript-eslint/eslint-plugin": "^8.0.0-alpha.18",
-    "@typescript-eslint/parser": "^8.0.0-alpha.18",
-    "eslint": "^9.3.0",
+    "@typescript-eslint/eslint-plugin": "^8.7.0",
+    "@typescript-eslint/parser": "^8.7.0",
+    "eslint": "^9.11.1",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import-x": "^0.5.1",
-    "eslint-plugin-svelte": "^2.39.0",
-    "eslint-plugin-unicorn": "^53.0.0",
+    "eslint-plugin-import-x": "^4.3.0",
+    "eslint-plugin-svelte": "^2.44.0",
+    "eslint-plugin-unicorn": "^55.0.0",
     "estree-walker": "^3.0.3",
-    "globals": "^15.3.0",
-    "magic-string": "^0.30.10",
-    "material-symbols": "^0.18.0",
-    "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^3.2.3",
-    "publint": "^0.2.8",
-    "sass": "^1.77.4",
-    "svelte": "5.0.0-next.142",
-    "svelte-check": "^3.7.1",
-    "svelte-eslint-parser": "^0.36.0",
+    "globals": "^15.9.0",
+    "magic-string": "^0.30.11",
+    "material-symbols": "^0.23.0",
+    "prettier": "^3.3.3",
+    "prettier-plugin-svelte": "^3.2.6",
+    "publint": "^0.2.11",
+    "sass": "^1.79.3",
+    "svelte": "5.0.0-next.257",
+    "svelte-check": "^4.0.3",
+    "svelte-eslint-parser": "^0.41.1",
     "ts-node": "^10.9.2",
-    "tslib": "^2.6.2",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.11",
-    "vitest": "^1.6.0"
+    "tslib": "^2.7.0",
+    "typescript": "^5.6.2",
+    "vite": "^5.4.7",
+    "vitest": "^2.1.1"
   },
   "svelte": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "type": "module",
   "dependencies": {
-    "highlight.js": "^11.9.0",
-    "svelte-highlight": "^7.6.1"
+    "highlight.js": "^11.10.0",
+    "svelte-highlight": "^7.7.0"
   }
 }
