@@ -22,53 +22,58 @@
   import type { QLayoutProps } from "$lib/components/layout/props";
   import { snippet } from "./docs.snippets";
 
-  let displayLeftDrawerElement: QDrawer;
-  let displayLeftDrawer = false;
-  let viewArr = [
+  let displayLeftDrawerElement = $state<ReturnType<typeof QDrawer>>();
+  let displayLeftDrawer = $state(false);
+  let viewArr = $state([
     ["h", "h", "h"],
     ["l", "p", "r"],
     ["f", "f", "f"],
-  ];
-  $: view = viewArr.map((v) => v.join("")).join(" ") as QLayoutProps["view"];
+  ]);
+  const view = $derived(viewArr.map((v) => v.join("")).join(" ") as QLayoutProps["view"]);
 
-  let header = true;
-  let footer = true;
-  let leftRailbar = true;
-  let leftDrawer = true;
-  let rightRailbar = false;
-  let rightDrawer = false;
+  let showHeader = $state(true);
+  let showFooter = $state(true);
+  let leftRailbar = $state(true);
+  let leftDrawer = $state(true);
+  let rightRailbar = $state(false);
+  let rightDrawer = $state(false);
 
-  $: snippets = snippet(view, [header, footer, leftRailbar, leftDrawer, rightRailbar, rightDrawer]);
+  const snippets = $derived(
+    snippet(view, [showHeader, showFooter, leftRailbar, leftDrawer, rightRailbar, rightDrawer])
+  );
 
-  $: style =
+  const style = $derived(
     createStyles({
-      "--header-height": header ? "64px" : "0px",
-      "--footer-height": footer ? "80px" : "0px",
+      "--header-height": showHeader ? "64px" : "0px",
+      "--footer-height": showFooter ? "80px" : "0px",
       "--left-railbar-width": leftRailbar ? "88px" : "0px",
       "--right-railbar-width": rightRailbar ? "88px" : "0px",
       "--left-drawer-width": leftDrawer ? "300px" : "0px",
       "--right-drawer-width": rightDrawer ? "300px" : "0px",
-    }) || undefined;
+    }) || undefined
+  );
 
-  let leftDrawerElement: QDrawer;
-  let leftDrawerShown = true;
+  let leftDrawerElement = $state<ReturnType<typeof QDrawer>>();
+  let leftDrawerShown = $state(true);
 
-  let rightDrawerElement: QDrawer;
-  let rightDrawerShown = true;
+  let rightDrawerElement = $state<ReturnType<typeof QDrawer>>();
+  let rightDrawerShown = $state(true);
 
-  $: if (!leftDrawer) {
-    leftDrawerShown = false;
-  } else if (!rightDrawer) {
-    rightDrawerShown = false;
-  }
+  $effect(() => {
+    if (!leftDrawer) {
+      leftDrawerShown = false;
+    } else if (!rightDrawer) {
+      rightDrawerShown = false;
+    }
+  });
 </script>
 
 <QDocs componentDocs={QLayoutDocs}>
   {#snippet display()}
-    <QLayout view="lhh lpr lfr" headerHeight="50px" footerHeight="50px">
+    <QLayout view="lhh lpr lfr" headerHeight="50px" footerHeight="50px" style="border-radius: 0">
       {#snippet header()}
-        <QHeader elevate>
-          <QBtn icon="menu" design="flat" onclick={displayLeftDrawerElement.toggle} />
+        <QHeader elevated height={48}>
+          <QBtn icon="menu" design="flat" onclick={displayLeftDrawerElement?.toggle} />
           <QToolbarTitle>Header</QToolbarTitle>
         </QHeader>
       {/snippet}
@@ -115,7 +120,7 @@
         </QRailbar>
       {/snippet}
       {#snippet footer()}
-        <QFooter class="no-round flex flex-center" elevate>
+        <QFooter class="no-round flex flex-center" height={48}>
           <h3 class="small center">Footer</h3>
         </QFooter>
       {/snippet}
@@ -128,177 +133,189 @@
         <QCard bordered class="q-pa-none" style="height: 80vh">
           <QLayout {view} {style}>
             {#snippet header()}
-              <QHeader style={header ? undefined : "display: none;"} elevate>
-                {#if leftDrawer}
-                  <QBtn icon="menu" design="flat" onclick={leftDrawerElement.toggle} />
-                {/if}
-                <div class="flex column">
-                  <QRadio bind:selected={viewArr[0][0]} value="h" label="h" />
-                  <QRadio style="margin: 0" bind:selected={viewArr[0][0]} value="l" label="l" />
-                </div>
-                <div class="max flex column items-center" style="flex-grow: 1">
-                  <QRadio
-                    style="width: fit-content"
-                    bind:selected={viewArr[0][1]}
-                    value="h"
-                    label="h"
-                  />
-                  <QRadio
-                    style="width: fit-content; margin: 0"
-                    bind:selected={viewArr[0][1]}
-                    value="H"
-                    label="H"
-                  />
-                </div>
-                <div class="flex column">
-                  <QRadio bind:selected={viewArr[0][2]} value="h" label="h" />
-                  <QRadio style="margin: 0" bind:selected={viewArr[0][2]} value="r" label="r" />
-                </div>
-                {#if rightDrawer}
-                  <QBtn icon="menu" design="flat" onclick={rightDrawerElement.toggle} />
-                {/if}
-              </QHeader>
+              {#if showHeader}
+                <QHeader elevated>
+                  {#if leftDrawer}
+                    <QBtn icon="menu" design="flat" onclick={leftDrawerElement?.toggle} />
+                  {/if}
+                  <div class="flex column">
+                    <QRadio bind:selected={viewArr[0][0]} value="h" label="h" />
+                    <QRadio style="margin: 0" bind:selected={viewArr[0][0]} value="l" label="l" />
+                  </div>
+                  <div class="max flex column items-center" style="flex-grow: 1">
+                    <QRadio
+                      style="width: fit-content"
+                      bind:selected={viewArr[0][1]}
+                      value="h"
+                      label="h"
+                    />
+                    <QRadio
+                      style="width: fit-content; margin: 0"
+                      bind:selected={viewArr[0][1]}
+                      value="H"
+                      label="H"
+                    />
+                  </div>
+                  <div class="flex column">
+                    <QRadio bind:selected={viewArr[0][2]} value="h" label="h" />
+                    <QRadio style="margin: 0" bind:selected={viewArr[0][2]} value="r" label="r" />
+                  </div>
+                  {#if rightDrawer}
+                    <QBtn icon="menu" design="flat" onclick={rightDrawerElement?.toggle} />
+                  {/if}
+                </QHeader>
+              {/if}
             {/snippet}
 
             {#snippet railbarLeft()}
-              <QRailbar style={leftRailbar ? undefined : "display: none;"} bordered>
-                <QList>
-                  <QItem to="#">
-                    <QIcon name="home" />
-                    <QItemSection>Home</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="help" />
-                    <QItemSection>About</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="shopping_cart" />
-                    <QItemSection>Store</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="mail" />
-                    <QItemSection>Contact</QItemSection>
-                  </QItem>
-                </QList>
-              </QRailbar>
+              {#if leftRailbar}
+                <QRailbar bordered>
+                  <QList>
+                    <QItem to="#">
+                      <QIcon name="home" />
+                      <QItemSection>Home</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="help" />
+                      <QItemSection>About</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="shopping_cart" />
+                      <QItemSection>Store</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="mail" />
+                      <QItemSection>Contact</QItemSection>
+                    </QItem>
+                  </QList>
+                </QRailbar>
+              {/if}
             {/snippet}
 
             {#snippet railbarRight()}
-              <QRailbar style={rightRailbar ? undefined : "display: none;"} side="right" bordered>
-                <QList>
-                  <QItem to="#">
-                    <QIcon name="home" />
-                    <QItemSection>Home</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="help" />
-                    <QItemSection>About</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="shopping_cart" />
-                    <QItemSection>Store</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="mail" />
-                    <QItemSection>Contact</QItemSection>
-                  </QItem>
-                </QList>
-              </QRailbar>
+              {#if rightRailbar}
+                <QRailbar side="right" bordered>
+                  <QList>
+                    <QItem to="#">
+                      <QIcon name="home" />
+                      <QItemSection>Home</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="help" />
+                      <QItemSection>About</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="shopping_cart" />
+                      <QItemSection>Store</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="mail" />
+                      <QItemSection>Contact</QItemSection>
+                    </QItem>
+                  </QList>
+                </QRailbar>
+              {/if}
             {/snippet}
             {#snippet drawerLeft()}
-              <QDrawer
-                style={leftDrawer ? undefined : "display: none;"}
-                bordered
-                persistent
-                bind:value={leftDrawerShown}
-                bind:this={leftDrawerElement}
-              >
-                <QList>
-                  <QItem to="#">
-                    <QIcon name="home" />
-                    <QItemSection>Home</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="help" />
-                    <QItemSection>About</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="shopping_cart" />
-                    <QItemSection>Store</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="mail" />
-                    <QItemSection>Contact</QItemSection>
-                  </QItem>
-                </QList>
-              </QDrawer>
+              {#if leftDrawer}
+                <QDrawer
+                  bordered
+                  width={160}
+                  persistent
+                  bind:value={leftDrawerShown}
+                  bind:this={leftDrawerElement}
+                >
+                  <QList>
+                    <QItem to="#">
+                      <QIcon name="home" />
+                      <QItemSection>Home</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="help" />
+                      <QItemSection>About</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="shopping_cart" />
+                      <QItemSection>Store</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="mail" />
+                      <QItemSection>Contact</QItemSection>
+                    </QItem>
+                  </QList>
+                </QDrawer>
+              {/if}
             {/snippet}
             {#snippet drawerRight()}
-              <QDrawer
-                style={rightDrawer ? undefined : "display: none;"}
-                side="right"
-                bordered
-                persistent
-                bind:value={rightDrawerShown}
-                bind:this={rightDrawerElement}
-              >
-                <QList>
-                  <QItem to="#">
-                    <QIcon name="home" />
-                    <QItemSection>Home</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="help" />
-                    <QItemSection>About</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="shopping_cart" />
-                    <QItemSection>Store</QItemSection>
-                  </QItem>
-                  <QItem to="#">
-                    <QIcon name="mail" />
-                    <QItemSection>Contact</QItemSection>
-                  </QItem>
-                </QList>
-              </QDrawer>
+              {#if rightDrawer}
+                <QDrawer
+                  side="right"
+                  bordered
+                  persistent
+                  width={160}
+                  bind:value={rightDrawerShown}
+                  bind:this={rightDrawerElement}
+                >
+                  <QList>
+                    <QItem to="#">
+                      <QIcon name="home" />
+                      <QItemSection>Home</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="help" />
+                      <QItemSection>About</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="shopping_cart" />
+                      <QItemSection>Store</QItemSection>
+                    </QItem>
+                    <QItem to="#">
+                      <QIcon name="mail" />
+                      <QItemSection>Contact</QItemSection>
+                    </QItem>
+                  </QList>
+                </QDrawer>
+              {/if}
             {/snippet}
             {#snippet footer()}
-              <QFooter style={footer ? undefined : "display: none;"} elevate>
-                {#if leftDrawer}
-                  <QBtn icon="menu" design="flat" onclick={leftDrawerElement.toggle} />
-                {/if}
-                <div class="flex column">
-                  <QRadio bind:selected={viewArr[2][0]} value="f" label="f" />
-                  <QRadio style="margin: 0" bind:selected={viewArr[2][0]} value="l" label="l" />
-                </div>
-                <div class="max flex column items-center" style="flex-grow: 1">
-                  <QRadio
-                    style="width: fit-content"
-                    bind:selected={viewArr[2][1]}
-                    value="f"
-                    label="f"
-                  />
-                  <QRadio
-                    style="width: fit-content; margin: 0"
-                    bind:selected={viewArr[2][1]}
-                    value="F"
-                    label="F"
-                  />
-                </div>
-                <div class="flex column">
-                  <QRadio bind:selected={viewArr[2][2]} value="f" label="f" />
-                  <QRadio style="margin: 0" bind:selected={viewArr[2][2]} value="r" label="r" />
-                </div>
-                {#if rightDrawer}
-                  <QBtn icon="menu" design="flat" onclick={rightDrawerElement.toggle} />
-                {/if}
-              </QFooter>
+              {#if showFooter}
+                <QFooter bordered>
+                  {#if leftDrawer}
+                    <QBtn icon="menu" design="flat" onclick={leftDrawerElement?.toggle} />
+                  {/if}
+                  <div class="flex column">
+                    <QRadio bind:selected={viewArr[2][0]} value="f" label="f" />
+                    <QRadio style="margin: 0" bind:selected={viewArr[2][0]} value="l" label="l" />
+                  </div>
+                  <div class="max flex column items-center" style="flex-grow: 1">
+                    <QRadio
+                      style="width: fit-content"
+                      bind:selected={viewArr[2][1]}
+                      value="f"
+                      label="f"
+                    />
+                    <QRadio
+                      style="width: fit-content; margin: 0"
+                      bind:selected={viewArr[2][1]}
+                      value="F"
+                      label="F"
+                    />
+                  </div>
+                  <div class="flex column">
+                    <QRadio bind:selected={viewArr[2][2]} value="f" label="f" />
+                    <QRadio style="margin: 0" bind:selected={viewArr[2][2]} value="r" label="r" />
+                  </div>
+                  {#if rightDrawer}
+                    <QBtn icon="menu" design="flat" onclick={rightDrawerElement?.toggle} />
+                  {/if}
+                </QFooter>
+              {/if}
             {/snippet}
             {#snippet content()}
-              <QList>
+              <QList style="text-align: center">
                 <QItem>
                   <QItemSection type="toggle">
-                    <QSwitch bind:value={header} />
+                    <QSwitch bind:value={showHeader} />
                   </QItemSection>
                   <QItemSection>Header</QItemSection>
                 </QItem>
@@ -328,7 +345,7 @@
                 </QItem>
                 <QItem>
                   <QItemSection type="toggle">
-                    <QSwitch bind:value={footer} />
+                    <QSwitch bind:value={showFooter} />
                   </QItemSection>
                   <QItemSection>Footer</QItemSection>
                 </QItem>
@@ -340,3 +357,14 @@
     </div>
   {/snippet}
 </QDocs>
+
+<style lang="scss">
+  :global(.q-layout__content > .q-list) {
+    max-width: fit-content;
+    margin-inline: auto;
+  }
+
+  :global(.q-footer > nav) {
+    justify-content: center;
+  }
+</style>
