@@ -1,9 +1,11 @@
 <script lang="ts">
+  import { getContext, onDestroy, untrack } from "svelte";
   import { navigating } from "$app/stores";
   import { useSize } from "$lib/composables";
   import { clickOutside } from "$lib/helpers";
   import QContext from "$lib/classes/QContext.svelte";
-  import type { LayoutContext } from "../layout/QLayout.svelte";
+  import type { QLayoutProps } from "$components/layout/props";
+  import type { DrawerContext, LayoutContext } from "../layout/QLayout.svelte";
   import type { QDrawerProps } from "./props";
 
   let {
@@ -17,6 +19,9 @@
     ...props
   }: QDrawerProps = $props();
 
+  const drawerContext = QContext.get<DrawerContext>(`QDrawer-${side}`);
+  const layoutView = getContext<{ value: NonNullable<QLayoutProps["view"]> }>("view");
+
   const ctx = QContext.get<LayoutContext>("layout");
 
   const drawerType = $derived(side === "left" ? "drawerLeft" : "drawerRight");
@@ -26,6 +31,15 @@
   const canHideOnClickOutside = $derived((value && !persistent) || overlay);
 
   const hideOnRouteChange = $derived(!persistent || overlay);
+
+  const offsetTop = $derived.by(() => {
+    const charPos = side === "left" ? 0 : 2;
+    return layoutView?.value.charAt(charPos) === "h";
+  });
+  const offsetBottom = $derived.by(() => {
+    const charPos = side === "left" ? 8 : 10;
+    return layoutView?.value.charAt(charPos) === "f";
+  });
 
   export const show = (e?: MouseEvent) => {
     if (!value) {
@@ -52,6 +66,27 @@
     }
   });
 
+  onDestroy(() => {
+    drawerContext?.updateEntries({
+      width: 0,
+      takesSpace: false,
+    });
+  });
+
+  $effect(() => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-expressions
+    [value, overlay, width];
+
+    // TODO: Try to make it work without console.debug
+    // writing untrack(() => drawerContext)?.updateEntries(...) doesn't work (infinite $effect loop)
+    untrack(() =>
+      drawerContext?.updateEntries({
+        takesSpace: !!value && !overlay,
+        width,
+      })
+    );
+  });
+
   const shouldHaveRadius = (pos: "top" | "bottom") => {
     const appBarEl = pos === "top" ? ctx?.value?.header : ctx?.value?.footer;
 
@@ -65,14 +100,13 @@
       active: value,
       overlay,
       bordered,
-      "offset-top": drawerCtx?.offset.top,
-      "offset-bottom": drawerCtx?.offset.bottom,
+      "offset-top": offsetTop,
+      "offset-bottom": offsetBottom,
       fixed: drawerCtx?.fixed,
       "top-left-radius": side === "right" && shouldHaveRadius("top"),
       "bottom-left-radius": side === "right" && shouldHaveRadius("bottom"),
       "top-right-radius": side === "left" && shouldHaveRadius("top"),
       "bottom-right-radius": side === "left" && shouldHaveRadius("bottom"),
-      above: (["top", "bottom"] as const).some((pos) => !drawerCtx?.offset[pos] && overlay),
     },
     classes: [props.class],
   });
