@@ -1,36 +1,76 @@
 <script lang="ts">
-  import { useSize } from "$lib/composables";
+  import { getContext, onDestroy, onMount, untrack } from "svelte";
   import QContext from "$lib/classes/QContext.svelte";
-  import type { LayoutContext } from "../layout/QLayout.svelte";
+  import QScrollObserver from "$lib/classes/QScrollObserver.svelte";
+  import type { QLayoutProps } from "$components/layout/props";
+  import QToolbar from "$components/toolbar/QToolbar.svelte";
+  import type { AppbarContext } from "../layout/QLayout.svelte";
   import type { QFooterProps } from "./props";
+
+  const footerIdentifier = Date.now();
 
   let {
     value = $bindable(true),
-    border = false,
-    elevate = false,
-    height,
+    bordered = false,
+    reveal = false,
+    revealOffset = 250,
+    height = 80,
     children,
     ...props
   }: QFooterProps = $props();
 
-  const ctx = QContext.get<LayoutContext>("layout");
+  const footerContext = QContext.get<AppbarContext>("QFooter");
+  const layoutView = getContext<{ value: NonNullable<QLayoutProps["view"]> }>("view");
+
+  if (!footerContext || !layoutView) {
+    throw new Error("QFooter should be used inside QLayout");
+  }
+
+  const scroll = $derived(
+    reveal ? new QScrollObserver(`.q-footer--${footerIdentifier} ~ .q-layout__content`) : undefined
+  );
+  let contentScrollHeight = $state(0);
+
+  const offset = $derived(scroll ? scroll.position + height : undefined);
+  // Collapse the footer `${reavealOffset}px` above the bottom of layout content when scrolling up
+  const collapsed = $derived(
+    reveal && scroll?.direction === "up" && offset! + revealOffset < contentScrollHeight
+  );
+
+  const leftOffset = () => layoutView.value.charAt(8) === "l";
+  const rightOffset = () => layoutView.value.charAt(10) === "r";
+
+  $effect.pre(() => {
+    untrack(() => footerContext).updateEntries({ height, collapsed });
+  });
+
+  onMount(() => {
+    // Calculating the layout content's height
+    const content = document.querySelector(`.q-footer--${footerIdentifier} ~ .q-layout__content`);
+
+    contentScrollHeight = content ? content.scrollHeight - content.clientHeight : 0;
+  });
+
+  onDestroy(() => {
+    untrack(() => footerContext).updateEntries({ height: 0, collapsed: false });
+  });
 
   Q.classes("q-footer", {
     bemClasses: {
-      bordered: border,
-      elevated: elevate,
-      fixed: ctx?.value?.footer?.fixed,
+      [footerIdentifier]: true,
+      collapsed,
+      bordered,
+      "offset-left": leftOffset(),
+      "offset-right": rightOffset(),
     },
     classes: [props.class],
   });
-
-  const heightStyle = $derived(!ctx ? useSize(height).style : null);
 </script>
 
 {#if value}
-  <footer {...props} class="q-footer" {...Q.classes} style:height={heightStyle}>
-    <nav>
+  <footer {...props} class="q-footer" {...Q.classes} style:--footer-height="{height}px">
+    <QToolbar>
       {@render children?.()}
-    </nav>
+    </QToolbar>
   </footer>
 {/if}
