@@ -1,53 +1,63 @@
-<svelte:options runes={true} />
-
 <script lang="ts">
-  import { untrack } from "svelte";
-  import { createClasses } from "$lib/utils";
-  import { QToolbar } from "$lib";
+  import { getContext, onDestroy, untrack } from "svelte";
+  import QToolbar from "$components/toolbar/QToolbar.svelte";
   import QContext from "$lib/classes/QContext.svelte";
-  import type { LayoutContext } from "../layout/QLayout.svelte";
+  import QScrollObserver from "$lib/classes/QScrollObserver.svelte";
+  import type { AppbarContext } from "$components/layout/QLayout.svelte";
+  import type { QLayoutProps } from "$components/layout/props";
   import type { QHeaderProps } from "./props";
 
+  const headerIdentifier = Date.now();
+
   let {
+    elevated = false,
     inset = false,
-    elevate = false,
-    border = false,
+    reveal = false,
+    revealOffset = 250,
+    height = 64,
     children,
     ...props
   }: QHeaderProps = $props();
 
-  const ctx = QContext.get<LayoutContext>("layout");
+  const headerContext = QContext.get<AppbarContext>("QHeader");
+  const layoutView = getContext<{ value: NonNullable<QLayoutProps["view"]> }>("view");
+  if (!headerContext || !layoutView) {
+    throw new Error("QHeader should be used inside QLayout");
+  }
 
-  const header = $derived(ctx?.value?.header);
+  const scroll = $derived(
+    reveal ? new QScrollObserver(`.q-header--${headerIdentifier} ~ .q-layout__content`) : undefined
+  );
+  const offset = $derived(scroll ? scroll.position - height : undefined);
+  // Collapse the header `${reavealOffset}px` below the top of layout content when scrolling down
+  const collapsed = $derived(reveal && scroll?.direction === "down" && offset! - revealOffset > 0);
 
-  $effect(() => {
-    if (!ctx) {
-      console.warn("QHeader should be used inside QLayout");
-    }
+  const leftOffset = () => layoutView.value.charAt(0) === "l";
+  const rightOffset = () => layoutView.value.charAt(2) === "r";
+
+  $effect.pre(() => {
+    untrack(() => headerContext).updateEntries({ height, collapsed });
   });
 
-  const classes = $derived(
-    createClasses([header?.fixed && "fixed"], {
-      component: "q-header",
-      userClasses: props.class,
-    })
-  );
+  onDestroy(() => {
+    untrack(() => headerContext).updateEntries({ height: 0, collapsed: false });
+  });
 
-  $effect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-expressions
-    props.style;
-
-    untrack(() => {
-      if (header) {
-        ctx?.updateEntry("header", {
-          ...header,
-          display: !props.style?.includes("display: none"),
-        });
-      }
-    });
+  Q.classes("q-header", {
+    bemClasses: {
+      [headerIdentifier]: true,
+      elevated,
+      collapsed,
+      "offset-left": leftOffset(),
+      "offset-right": rightOffset(),
+    },
+    classes: [props.class],
+    isCustomComponent: true,
   });
 </script>
 
-<QToolbar {...props} {inset} {elevate} {border} class={classes} role="header">
-  {@render children?.()}
-</QToolbar>
+<header {...props} class="q-header" {...Q.classes} style:--header-height="{height}px">
+  <QToolbar {inset}>
+    {@render children?.()}
+  </QToolbar>
+</header>
