@@ -112,6 +112,8 @@ export function shouldReduceMotion() {
   return window.matchMedia("(prefers-reduced-motion: reduce)").matches;
 }
 
-export function elFromSelector(selector: string | HTMLElement | null) {
-  return typeof selector === "string" ? document.querySelector<HTMLElement>(selector) : selector;
+export function elFromSelector<T extends HTMLElement>(
+  selector: string | T | null | typeof window | Document
+) {
+  return typeof selector === "string" ? document.querySelector<T>(selector) : selector;
 }
